+++ tests/unit/test_main.py
@@ -18,6 +18,37 @@ class TestMain:
         self.napp.controller.switches = {"00:00:00:00:00:00:00:01"}
         self.base_endpoint = "kytos/of_multi_table/v1"
 
+    def test_default_pipeline(self) -> None:
+        """Test default pipeline value.
+
+        This is to ensure that if settings.DEFAULT_PIPELINE changes
+        then it needs to be updated accordingly.
+        """
+        assert self.napp.default_pipeline == {
+            "multi_table": [
+                {
+                    "table_id": 0,
+                    "napps_table_groups": {
+                        "coloring": ["base"],
+                        "of_lldp": ["base"],
+                        "mef_eline": ["evpl", "epl"],
+                    },
+                },
+                {
+                    "table_id": 2,
+                    "napps_table_groups": {
+                        "telemetry_int": ["evpl"],
+                    },
+                },
+                {
+                    "table_id": 3,
+                    "napps_table_groups": {
+                        "telemetry_int": ["epl"],
+                    },
+                },
+            ]
+        }
+
     async def test_get_enabled_table(self):
         """Test get the enabled table"""
         controller = self.napp.pipeline_controller