+++ CHANGELOG.rst
@@ -1,10 +1,14 @@
 [UNRELEASED] - Under development
 ********************************
 
+Added
+=====
+- Subscribed to ``"kytos/telemetry_int.enable_table"`` to support ``telemetry_int``
+
 Changed
 =======
 
-- Replaced ``telemetry_int`` ``base`` table group with ``epl`` and ``evpl``
+- Replaced ``telemetry_int`` ``base`` table group with ``evpl`` and ``epl`` by default using table 2 and 3 respectively
 
 [2023.1.0] - 2023-06-12
 ***********************